requirements.txt
streamlit
pandas
plotly
watchdog
yfinance
python-dotenv
